requirements.dev.txt
flask 
flask_cors
sqlitedict
request
pickle